# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OdontoFlow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/availability-rules": {
      "post": {
        "summary": "Create Availability Rule Route",
        "operationId": "create_availability_rule_route_availability_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityRuleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedule-blocks": {
      "post": {
        "summary": "Create Schedule Block Route",
        "operationId": "create_schedule_block_route_schedule_blocks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBlockCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleBlockRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/slots/query": {
      "post": {
        "summary": "Query Slots Route",
        "operationId": "query_slots_route_slots_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SlotResult"
                  },
                  "type": "array",
                  "title": "Response Query Slots Route Slots Query Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments": {
      "get": {
        "summary": "List Appointments Route",
        "description": "Agenda read: the acting organization's appointments, half-open window.\n\nThe window is ``[from_date, to_date)``; both bounds are optional and\npassed as RFC 3339 instants. Location/practitioner filters narrow the\nresult; the tenant always comes from the context, never from a query\nparameter (X3).",
        "operationId": "list_appointments_route_appointments_get",
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "practitioner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Practitioner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentListItem"
                  },
                  "title": "Response List Appointments Route Appointments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Appointment Route",
        "operationId": "create_appointment_route_appointments_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{appointment_id}": {
      "get": {
        "summary": "Get Appointment Route",
        "operationId": "get_appointment_route_appointments__appointment_id__get",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{appointment_id}/cancel": {
      "post": {
        "summary": "Cancel Appointment Route",
        "operationId": "cancel_appointment_route_appointments__appointment_id__cancel_post",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AppointmentCancel"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{appointment_id}/reschedule": {
      "post": {
        "summary": "Reschedule Appointment Route",
        "operationId": "reschedule_appointment_route_appointments__appointment_id__reschedule_post",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentReschedule"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AppointmentCancel": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "AppointmentCancel",
        "description": "Empty by design: the appointment is identified by the path, and nothing\nabout the cancellation is caller-supplied. ``extra='forbid'`` keeps a\nclient from smuggling state through the body."
      },
      "AppointmentCreate": {
        "properties": {
          "lead_id": {
            "type": "integer",
            "title": "Lead Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lead_id",
          "service_id",
          "location_id",
          "practitioner_id",
          "start"
        ],
        "title": "AppointmentCreate"
      },
      "AppointmentListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lead_id": {
            "type": "integer",
            "title": "Lead Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "start_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Start Utc"
          },
          "end_utc": {
            "type": "string",
            "format": "date-time",
            "title": "End Utc"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "lead_name": {
            "type": "string",
            "title": "Lead Name"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "practitioner_name": {
            "type": "string",
            "title": "Practitioner Name"
          },
          "location_name": {
            "type": "string",
            "title": "Location Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lead_id",
          "service_id",
          "practitioner_id",
          "location_id",
          "start_utc",
          "end_utc",
          "state",
          "lead_name",
          "service_name",
          "practitioner_name",
          "location_name"
        ],
        "title": "AppointmentListItem",
        "description": "Agenda read DTO: the appointment plus the names the agenda renders.\n\nThe names are joined at the application boundary (relationships), so the\nfrontend never needs per-row lookups and never sees raw mockData shapes."
      },
      "AppointmentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lead_id": {
            "type": "integer",
            "title": "Lead Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "start_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Start Utc"
          },
          "end_utc": {
            "type": "string",
            "format": "date-time",
            "title": "End Utc"
          },
          "state": {
            "type": "string",
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lead_id",
          "service_id",
          "practitioner_id",
          "location_id",
          "start_utc",
          "end_utc",
          "state"
        ],
        "title": "AppointmentRead"
      },
      "AppointmentReschedule": {
        "properties": {
          "new_start": {
            "type": "string",
            "format": "date-time",
            "title": "New Start"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "new_start"
        ],
        "title": "AppointmentReschedule"
      },
      "AvailabilityRuleCreate": {
        "properties": {
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "day_of_week": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Day Of Week"
          },
          "start_local": {
            "type": "string",
            "format": "time",
            "title": "Start Local"
          },
          "end_local": {
            "type": "string",
            "format": "time",
            "title": "End Local"
          }
        },
        "type": "object",
        "required": [
          "practitioner_id",
          "location_id",
          "day_of_week",
          "start_local",
          "end_local"
        ],
        "title": "AvailabilityRuleCreate"
      },
      "AvailabilityRuleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "start_local": {
            "type": "string",
            "format": "time",
            "title": "Start Local"
          },
          "end_local": {
            "type": "string",
            "format": "time",
            "title": "End Local"
          }
        },
        "type": "object",
        "required": [
          "id",
          "practitioner_id",
          "location_id",
          "day_of_week",
          "start_local",
          "end_local"
        ],
        "title": "AvailabilityRuleRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleBlockCreate": {
        "properties": {
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "start_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Start Utc"
          },
          "end_utc": {
            "type": "string",
            "format": "date-time",
            "title": "End Utc"
          }
        },
        "type": "object",
        "required": [
          "practitioner_id",
          "location_id",
          "start_utc",
          "end_utc"
        ],
        "title": "ScheduleBlockCreate"
      },
      "ScheduleBlockRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "start_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Start Utc"
          },
          "end_utc": {
            "type": "string",
            "format": "date-time",
            "title": "End Utc"
          }
        },
        "type": "object",
        "required": [
          "id",
          "practitioner_id",
          "location_id",
          "start_utc",
          "end_utc"
        ],
        "title": "ScheduleBlockRead"
      },
...
      "SlotQuery": {
        "properties": {
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          }
        },
        "type": "object",
        "required": [
          "service_id",
          "location_id",
          "window_start",
          "window_end"
        ],
        "title": "SlotQuery"
      },
      "SlotResult": {
        "properties": {
          "practitioner_id": {
            "type": "integer",
            "title": "Practitioner Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "practitioner_id",
          "start",
          "end"
        ],
        "title": "SlotResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}